requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.35.0
pydantic==2.5.0
python-dotenv==1.0.0

# Google Cloud and Vertex AI
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-auth==2.23.4

# AI/ML and LangChain
langchain==0.1.0
langchain-google-vertexai==0.0.1
langchain-community==0.0.10
vertexai==1.38.1

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# Document processing
pypdf==3.17.1
pymupdf==1.23.8
python-multipart==0.0.6

# HTTP and networking
httpx==0.25.2
websockets==15.0.1
websocket-client==1.8.0

# Data processing
pandas==2.1.4
numpy==1.25.2

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: For enhanced functionality
# openai==1.3.7  # If using OpenAI models
# anthropic==0.7.8  # If using Anthropic models 